backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.2.1
SQLAlchemy==2.0.32
alembic==1.13.2
python-multipart==0.0.9
pdfminer.six==20231228
numpy==1.26.4
openai==1.12.0
sentence-transformers==2.5.1
scikit-learn==1.4.0
pandas==2.2.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
pgvector==0.2.4